Replace XdnGeoDemandProfiler with XdnGeoDemandProfiler2
