+++ conf/gigapaxos.xdn.local.geodemand.properties
@@ -26,7 +26,7 @@ HTTP_AR_FRONTEND_BATCH_ENABLED=true
 HTTP_AR_FRONTEND_REQUEST_TIMEOUT_MS=60000
 XDN_HTTP_MAX_POOL_SIZE=128
 
-DEMAND_PROFILE_TYPE=edu.umass.cs.xdn.XdnGeoDemandProfiler2
+DEMAND_PROFILE_TYPE=edu.umass.cs.xdn.XdnGeoDemandProfiler
 XDN_GEO_NUM_GRID_ROWS=800
 XDN_GEO_NUM_GRID_COLUMNS=400
 XDN_GEO_PLACEMENT_ALGORITHM=GREEDY
