+++ conf/logging.properties
@@ -38,7 +38,7 @@ java.util.logging.ConsoleHandler.level=FINEST
 edu.umass.cs.level=WARNING
 edu.umass.cs.xdn.XdnGigapaxosApp.level=WARNING
 edu.umass.cs.xdn.XdnGeoDemandProfiler.level=FINEST
-edu.umass.cs.xdn.XdnGeoDemandProfiler2.level=FINEST
+edu.umass.cs.xdn.XdnGeoDemandProfiler.level=FINEST
 edu.umass.cs.xdn.utils.GeoIpLookup.level=WARNING
 edu.umass.cs.xdn.XdnReplicaCoordinator.level=WARNING
 edu.umass.cs.xdn.XdnHttpRequestBatcher.level=WARNING
